requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
click==7.1.2
Django==3.1.1
django-debug-toolbar==2.2
django-lineage==0.2.0
django-multiselectfield==0.1.12
flake8==3.8.3
importlib-metadata==1.7.0
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.1.0
